parallel batch run all nodes to completion
Switching from `Promise.all()` to `Promise.allSettled()`. This ensures
that all nodes in the batch are run to completion, regardless of
individual failures. You can then inspect the results to see which
succeeded and which failed, allowing you to log all errors or even retry
only the failed items (TODO).